+++ core/api/src/main/java/app/kaito_dogi/mybrary/core/api/OkHttpInterceptor.kt
@@ -1,4 +1,4 @@
-package app.kaito_dogi.mybrary.core.api.okhttp
+package app.kaito_dogi.mybrary.core.api
 
 import javax.inject.Qualifier
