+++ core/api/src/main/java/app/kaito_dogi/mybrary/core/api/googlebooks/response/GetVolumeResponse.kt
@@ -1,0 +1,5 @@
+package app.kaito_dogi.mybrary.core.api.googlebooks.response
+
+import app.kaito_dogi.mybrary.core.api.googlebooks.response.model.ItemResponse
+
+typealias GetVolumeResponse = ItemResponse
