--- core/api/src/main/java/app/kaito_dogi/mybrary/core/api/response/GetVolumeResponse.kt
@@ -1,5 +1,0 @@
-package app.kaito_dogi.mybrary.core.api.response
-
-import app.kaito_dogi.mybrary.core.api.response.model.ItemResponse
-
-typealias GetVolumeResponse = ItemResponse
